Feature/jng 2073 generate spacers for layout

+++ judo-tatami-esm2ui/src/test/java/hu/blackbelt/judo/tatami/esm2ui/Northwind2UiApplicationTest.java
@@ -135,8 +135,8 @@ public class Northwind2UiApplicationTest {
         
         if(UiUtils.isHorizontalFlex(element)) {
           sb.append("---HORIZONTAL---");
-          int sumCol = UiUtils.getSumOfCols((Container) element);
-          int maxRow = UiUtils.getMaxOfRows((Container) element);
+          double sumCol = UiUtils.getSumOfCols((Container) element);
+          double maxRow = UiUtils.getMaxOfRows((Container) element);
           sb.append(maxRow + " rows in " + element.getRow());
           if (maxRow <= element.getRow()) sb.append(" ***FITS*** //////// ");
           else sb.append(" ===WRONG=== //////// ");
@@ -146,8 +146,8 @@ public class Northwind2UiApplicationTest {
         }
         else if(UiUtils.isVerticalFlex(element)) {
           sb.append("|||VERTICAL||| ");
-          int sumRow = UiUtils.getSumOfRows((Container) element);
-          int maxCol = UiUtils.getMaxOfCols((Container) element);
+          double sumRow = UiUtils.getSumOfRows((Container) element);
+          double maxCol = UiUtils.getMaxOfCols((Container) element);
 
           sb.append(sumRow + " rows in " + element.getRow());
           if (sumRow <= element.getRow()) sb.append(" ***FITS*** //////// ");
